requirements.txt
Flask>=2.0
flask-cors>=3.0
python-dotenv>=0.21
requests>=2.28
